+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `Direction` a struct holding a direction in azimuth and elevation
 - `DelayAndSum::expected_len()`
+- `Audio::from_pcm_data()`
 
 ### Changed
 - **BREAKING CHANGE** replaced many uses of `(azimuth, elevation)` with new `Direction` struct